fixed category & refactor the code structure and build

+++ src/Deed/Deed.Application/Capitals/Responses/CapitalResponse.cs
@@ -3,14 +3,14 @@ namespace Deed.Application.Capitals.Responses;
 public sealed record CapitalResponse(
     int Id,
     string Name,
-    float Balance,
+    decimal Balance,
     string Currency,
     bool IncludeInTotal,
     bool OnlyForSavings,
-    float TotalIncome,
-    float TotalExpense,
-    float TotalTransferIn,
-    float TotalTransferOut,
+    decimal TotalIncome,
+    decimal TotalExpense,
+    decimal TotalTransferIn,
+    decimal TotalTransferOut,
     DateTimeOffset CreatedAt,
     int CreatedBy
 );
