+++ src/Deed/Deed.Application/Categories/Commands/Update/UpdateCategoryCommand.cs
@@ -6,7 +6,7 @@ namespace Deed.Application.Categories.Commands.Update;
 public sealed record UpdateCategoryCommand(
     int Id,
     string? Name = null,
-    float? PlannedPeriodAmount = null,
+    decimal? PlannedPeriodAmount = null,
     PerPeriodType? Period = null,
     CategoryType? Type = null)
     : ICommand;
