+++ src/Deed/Deed.Application/Expenses/Requests/UpdateExpenseRequest.cs
@@ -3,6 +3,6 @@ namespace Deed.Application.Expenses.Requests;
 public sealed record UpdateExpenseRequest(
     int Id,
     int? CategoryId,
-    float? Amount,
+    decimal? Amount,
     string? Purpose,
     DateTimeOffset? Date);
