+++ src/Deed/Deed.Application/Expenses/Responses/CategoryExpense.cs
@@ -10,9 +10,9 @@ namespace Deed.Application.Expenses.Responses;
 public sealed record CategoryExpenseResponse(
     int CategoryId,
     string Name,
-    float CategorySum,
-    float Percentage,
-    float PlannedPeriodAmount,
+    decimal CategorySum,
+    decimal Percentage,
+    decimal PlannedPeriodAmount,
     string PeriodType,
     IEnumerable<ExpenseResponse> Expenses
 );
