+++ src/Deed/Deed.Domain/Entities/Expense.cs
@@ -12,7 +12,7 @@ public sealed class Expense
     {
     }
 
-    public required float Amount { get; set; }
+    public required decimal Amount { get; set; }
 
     public required DateTimeOffset PaymentDate { get; set; }
