+++ src/Deed/Deed.Domain/Entities/Transfer.cs
@@ -3,7 +3,7 @@ namespace Deed.Domain.Entities;
 public sealed class Transfer
     : Entity, IAuditableEntity, ISoftDeletableEntity
 {
-    public required float Amount { get; set; }
+    public required decimal Amount { get; set; }
 
     public int? SourceCapitalId { get; init; }
